Introduce a WorkspaceStore for handling following
